Add iOS support via dart:ffi — dual engine (.litertlm FFI, .task MediaPipe)
- Build libLiteRtLm.dylib for ios_arm64 and ios_sim_arm64 via Bazel
- Build libStreamProxy.dylib for both iOS targets
- Add iOS to hook/build.dart with IOSSdk device/simulator detection
- Extract FfiInferenceModel from desktop into core/ffi/ (shared code)
- Add dual engine in FlutterGemmaMobile.createModel():
  .litertlm on iOS → FFI (LiteRtLmFfiClient)
  .task on iOS/Android → MediaPipe (existing Pigeon+Swift path)
- Add build_ios.sh build script
- flutter build ios --debug succeeds, dylibs in bundle
