+++ native/litert_lm/build_ios.sh
@@ -1,0 +1,128 @@
+#!/bin/bash
+# Build libLiteRtLm.dylib for iOS device (arm64) and simulator (arm64).
+#
+# Prerequisites:
+#   - Bazel (via bazelisk): brew install bazelisk
+#   - Xcode with iOS SDK
+#   - Git LFS: brew install git-lfs
+#
+# Usage:
+#   ./build_ios.sh [version]
+#   ./build_ios.sh v0.10.2
+#   ./build_ios.sh          # uses latest tag
+
+set -euo pipefail
+
+SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
+LITERT_LM_DIR="/tmp/LiteRT-LM"
+VERSION="${1:-}"
+
+echo "=== Building libLiteRtLm.dylib for iOS ==="
+
+# 1. Clone or update LiteRT-LM
+if [ -d "$LITERT_LM_DIR" ]; then
+  echo "Updating $LITERT_LM_DIR..."
+  cd "$LITERT_LM_DIR"
+  git fetch --tags
+else
+  echo "Cloning LiteRT-LM..."
+  git clone https://github.com/google-ai-edge/LiteRT-LM "$LITERT_LM_DIR"
+  cd "$LITERT_LM_DIR"
+fi
+
+# 2. Checkout version
+if [ -n "$VERSION" ]; then
+  echo "Checking out $VERSION..."
+  git checkout -f "$VERSION"
+else
+  LATEST_TAG=$(git tag -l "v*" | sort -V | tail -1)
+  echo "Checking out latest tag: $LATEST_TAG..."
+  git checkout -f "$LATEST_TAG"
+fi
+echo "Building from: $(git log --oneline -1)"
+
+# 3. Ensure shared library target exists
+if ! grep -q "linkshared" c/BUILD; then
+  echo "Adding shared library target to c/BUILD..."
+  cat >> c/BUILD << 'BUILDEOF'
+
+cc_binary(
+    name = "libLiteRtLm.dylib",
+    linkshared = True,
+    linkopts = select({
+        "@platforms//os:macos": ["-Wl,-exported_symbol,_LiteRt*", "-Wl,-exported_symbol,_litert_lm_*"],
+        "@platforms//os:ios": ["-Wl,-exported_symbol,_LiteRt*", "-Wl,-exported_symbol,_litert_lm_*"],
+        "@platforms//os:linux": ["-Wl,--export-dynamic-symbol=LiteRt*", "-Wl,--export-dynamic-symbol=litert_lm_*"],
+        "//conditions:default": [],
+    }),
+    visibility = ["//visibility:public"],
+    deps = [":engine"],
+)
+BUILDEOF
+fi
+
+# 4. Pull LFS files
+echo "Pulling LFS files..."
+git lfs pull --include="prebuilt/ios_arm64/*,prebuilt/ios_sim_arm64/*"
+
+# 5. Build for iOS device (arm64)
+echo ""
+echo "=== Building for iOS device (arm64) ==="
+bazelisk build --config=ios_arm64 '//c:libLiteRtLm.dylib'
+DEVICE_DIR="$SCRIPT_DIR/prebuilt/ios_arm64"
+mkdir -p "$DEVICE_DIR"
+cp bazel-bin/c/libLiteRtLm.dylib "$DEVICE_DIR/"
+chmod +w "$DEVICE_DIR/libLiteRtLm.dylib"
+install_name_tool -id @rpath/libLiteRtLm.dylib "$DEVICE_DIR/libLiteRtLm.dylib"
+install_name_tool -add_rpath '@loader_path/../../..' "$DEVICE_DIR/libLiteRtLm.dylib" 2>/dev/null || true
+
+# 6. Build for iOS simulator (arm64)
+echo ""
+echo "=== Building for iOS simulator (arm64) ==="
+bazelisk build --config=ios_sim_arm64 '//c:libLiteRtLm.dylib'
+SIM_DIR="$SCRIPT_DIR/prebuilt/ios_sim_arm64"
+mkdir -p "$SIM_DIR"
+cp bazel-bin/c/libLiteRtLm.dylib "$SIM_DIR/"
+chmod +w "$SIM_DIR/libLiteRtLm.dylib"
+install_name_tool -id @rpath/libLiteRtLm.dylib "$SIM_DIR/libLiteRtLm.dylib"
+install_name_tool -add_rpath '@loader_path/../../..' "$SIM_DIR/libLiteRtLm.dylib" 2>/dev/null || true
+
+# 7. Build StreamProxy for both targets
+echo ""
+echo "=== Building StreamProxy ==="
+clang -shared -o "$DEVICE_DIR/libStreamProxy.dylib" \
+  -arch arm64 -target arm64-apple-ios16.0 \
+  -isysroot "$(xcrun --sdk iphoneos --show-sdk-path)" \
+  -install_name @rpath/libStreamProxy.dylib \
+  -Wl,-headerpad_max_install_names \
+  "$SCRIPT_DIR/stream_proxy.c"
+echo "StreamProxy (device): OK"
+
+clang -shared -o "$SIM_DIR/libStreamProxy.dylib" \
+  -arch arm64 -target arm64-apple-ios16.0-simulator \
+  -isysroot "$(xcrun --sdk iphonesimulator --show-sdk-path)" \
+  -install_name @rpath/libStreamProxy.dylib \
+  -Wl,-headerpad_max_install_names \
+  "$SCRIPT_DIR/stream_proxy.c"
+echo "StreamProxy (simulator): OK"
+
+# 8. Copy companion libs
+echo ""
+echo "=== Copying companion libs ==="
+for lib in libGemmaModelConstraintProvider.dylib; do
+  [ -f "prebuilt/ios_arm64/$lib" ] && cp "prebuilt/ios_arm64/$lib" "$DEVICE_DIR/$lib" && echo "  $lib → device"
+  [ -f "prebuilt/ios_sim_arm64/$lib" ] && cp "prebuilt/ios_sim_arm64/$lib" "$SIM_DIR/$lib" && echo "  $lib → simulator"
+done
+
+# 9. Verify
+echo ""
+echo "=== Verification ==="
+echo "Device (ios_arm64):"
+ls -lh "$DEVICE_DIR/"
+nm -gU "$DEVICE_DIR/libLiteRtLm.dylib" | grep "litert_lm_engine_create" | head -1
+echo ""
+echo "Simulator (ios_sim_arm64):"
+ls -lh "$SIM_DIR/"
+nm -gU "$SIM_DIR/libLiteRtLm.dylib" | grep "litert_lm_engine_create" | head -1
+echo ""
+echo "=== Done ==="
